separate apollo adapter logic to a separate file
