+++ origin-growth/src/apollo/resolvers.js
@@ -3,6 +3,7 @@ const { GraphQLDateTime } = require('graphql-iso-date')
 //const db = require('./db')
 const { Fetcher } = require('../rules/rules')
 const { getLocationInfo } = require('../util/locationInfo')
+const { campaignToApolloObject } = require('./adapter')
 // const { GrowthInvite } = require('../resources/invite')
 
 // Resolvers define the technique for fetching the types in the schema.
@@ -28,7 +29,7 @@ const resolvers = {
       return {
         totalCount: campaigns.length,
         nodes: campaigns.map(
-          async campaign => await campaign.toApolloObject(args.walletAddress)
+          async campaign => await campaignToApolloObject(campaign, args.walletAddress)
         ),
         pageInfo: {
           endCursor: 'TODO implement',
